misc:  Fix some compile warnings.

+++ backend_shinkos1245.c
@@ -1004,7 +1004,7 @@ static int shinkos1245_read_parse(void *vctx, const void **vjob, int data_fd, in
 	}
 
 	/* Use larger of our copy counts */
-	if (job->jp.copies > copies)
+	if ((int)job->jp.copies > copies)
 		job->copies = job->jp.copies;
 	else
 		job->copies = copies;
